+++ Profile/Shopware6/Gateway/Connection/ConnectionFactoryInterface.php
@@ -8,10 +8,11 @@
 namespace SwagMigrationAssistant\Profile\Shopware6\Gateway\Connection;
 
 use Shopware\Core\Framework\Log\Package;
+use SwagMigrationAssistant\Migration\Gateway\HttpClientInterface;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
 
 #[Package('services-settings')]
 interface ConnectionFactoryInterface
 {
-    public function createApiClient(MigrationContextInterface $migrationContext): ?AuthClient;
+    public function createApiClient(MigrationContextInterface $migrationContext): ?HttpClientInterface;
 }