+++ Profile/Shopware6/Media/HttpMediaDownloadService.php
@@ -8,19 +8,32 @@
 namespace SwagMigrationAssistant\Profile\Shopware6\Media;
 
 use Shopware\Core\Framework\Log\Package;
+use SwagMigrationAssistant\Migration\DataSelection\DefaultEntities;
+use SwagMigrationAssistant\Migration\Gateway\HttpClientInterface;
+use SwagMigrationAssistant\Migration\Gateway\HttpSimpleClient;
+use SwagMigrationAssistant\Migration\Media\Processor\HttpDownloadServiceBase;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
 use SwagMigrationAssistant\Profile\Shopware\DataSelection\DataSet\MediaDataSet;
 use SwagMigrationAssistant\Profile\Shopware\Gateway\Api\ShopwareApiGateway;
-use SwagMigrationAssistant\Profile\Shopware\Media\HttpMediaDownloadService as BaseHttpMediaDownloadService;
 use SwagMigrationAssistant\Profile\Shopware6\Shopware6ProfileInterface;
 
 #[Package('services-settings')]
-class HttpMediaDownloadService extends BaseHttpMediaDownloadService
+class HttpMediaDownloadService extends HttpDownloadServiceBase
 {
     public function supports(MigrationContextInterface $migrationContext): bool
     {
         return $migrationContext->getProfile() instanceof Shopware6ProfileInterface
             && $migrationContext->getGateway()->getName() === ShopwareApiGateway::GATEWAY_NAME
             && $this->getDataSetEntity($migrationContext) === MediaDataSet::getEntity();
+    }
+
+    protected function getMediaEntity(): string
+    {
+        return DefaultEntities::MEDIA;
+    }
+
+    protected function getHttpClient(MigrationContextInterface $migrationContext): ?HttpClientInterface
+    {
+        return new HttpSimpleClient();
     }
 }